requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.12.4
pydantic-settings==2.10.1
python-multipart==0.0.20
qdrant-client==1.15.0
fastembed==0.7.3
cachetools==6.1.0
tenacity==9.1.2
huggingface-hub==0.34.4
structlog==25.4.0
pymupdf==1.26.4
pdfplumber==0.11.7
pandas==2.3.2
openpyxl==3.1.5
python-docx==1.2.0
llama-cpp-python==0.3.16
pytest==8.4.1
httpx==0.28.1
